+++ src/pages/Log-in-page/index.js
@@ -1,0 +1,1 @@
+export { default } from './Log-In-page';
